+++ apps/manifest-demo/webpack-host/webpack.config.js
@@ -1,6 +1,6 @@
 const path = require('path');
-const { registerPluginTSTranspiler } = require('nx/src/utils/nx-plugin.js');
-registerPluginTSTranspiler();
+// const { registerPluginTSTranspiler } = require('nx/src/utils/nx-plugin.js');
+// registerPluginTSTranspiler();
 const {
   ModuleFederationPlugin,
 } = require('@module-federation/enhanced/webpack');
@@ -20,6 +20,7 @@ module.exports = composePlugins(withNx(), withReact(), (config, context) => {
           'rspack_manifest_provider@http://localhost:3011/mf-manifest.json',
         'js-entry-provider':
           'rspack_js_entry_provider@http://localhost:3012/remoteEntry.js',
+        'modern-js-provider': 'app1@http://127.0.0.1:4001/mf-manifest.json',
       },
       filename: 'remoteEntry.js',
       exposes: {
@@ -28,10 +29,22 @@ module.exports = composePlugins(withNx(), withReact(), (config, context) => {
       shared: {
         lodash: {},
         antd: {},
-        react: {},
-        'react/': {},
-        'react-dom': {},
-        'react-dom/': {},
+        'react/': {
+          singleton: true,
+          requiredVersion: '^18.3.1',
+        },
+        react: {
+          singleton: true,
+          requiredVersion: '^18.3.1',
+        },
+        'react-dom': {
+          singleton: true,
+          requiredVersion: '^18.3.1',
+        },
+        'react-dom/': {
+          singleton: true,
+          requiredVersion: '^18.3.1',
+        },
       },
       experiments: { federationRuntime: 'hoisted' },
       runtimePlugins: [path.join(__dirname, './runtimePlugin.ts')],
@@ -56,10 +69,12 @@ module.exports = composePlugins(withNx(), withReact(), (config, context) => {
     scriptType: 'text/javascript',
   };
   config.optimization = {
+    ...config.optimization,
     runtimeChunk: 'single',
     minimize: false,
     moduleIds: 'named',
     chunkIds: 'named',
+    splitChunks: false,
   };
   config.output.publicPath = 'http://localhost:3013/';
   return config;